requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.62.3
idna==3.10
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.32.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pdf2image==1.16.3
pdfkit==1.0.0
pexels-api==1.0.1
Pillow==10.0.1
proto-plus==1.26.1
protobuf==4.25.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
six==1.17.0
smmap==5.0.2
streamlit==1.31.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
XlsxWriter==3.2.2
zipp==3.21.0